requirements.txt
absl-py==2.3.1
alembic==1.16.5
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
colorlog==6.9.0
contourpy==1.3.3
coverage==7.10.6
cupy-cuda12x==13.6.0
cycler==0.12.1
et_xmlfile==2.0.0
fastrlock==0.8.3
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.0
h5py==3.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kagglehub==0.3.13
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pandas-stubs==2.3.2.250827
pillow==11.3.0
pluggy==1.6.0
protobuf==6.32.1
Pygments==2.19.2
pyparsing==3.2.4
pytest==8.4.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
